Synchronizacja z klasą Account #Dziedziczenie
