+++ Bank/Printer.cs
@@ -8,24 +8,13 @@ namespace Bank
 {
     class Printer
     {
-        public void Print (SavingsAccount account)
+        public void Print (Account account)
         {
             Console.WriteLine($"Dane konta: {account.AccountNumber}");
-            Console.WriteLine($"Saldo: {account.Balance}");
-            Console.WriteLine($"Imię właściciela: {account.FirstName}");
-            Console.WriteLine($"Nazwisko właściciela: {account.LastName}");
-            Console.WriteLine($"PESEL właściciela: {account.Pesel}");
+            Console.WriteLine($"Typ konta: {account.TypeName()}");
             Console.WriteLine(account.GetBalance());
-            Console.WriteLine();
-        }
-        public void Print(BillingAccount account)
-        {
-            Console.WriteLine($"Dane konta: {account.AccountNumber}");
-            Console.WriteLine($"Saldo: {account.Balance}");
-            Console.WriteLine($"Imię właściciela: {account.FirstName}");
-            Console.WriteLine($"Nazwisko właściciela: {account.LastName}");
+            Console.WriteLine($"Imię i nazwisko właściciela: {account.GetFullName()}");
             Console.WriteLine($"PESEL właściciela: {account.Pesel}");
-            Console.WriteLine(account.GetBalance());
             Console.WriteLine();
         }
     }
